Complete SQLiteDatabaseManager and fix missing import
